some delete endpoint optimizations

+++ api/routes.py
@@ -284,7 +284,7 @@ def delete_photos(current_user: Token):
                 service.delete_user_photos_and_metadata(current_user, to_delete_user_id=user_id)
 
                 return "", 200
-            else:
+            elif current_app.config["IMG_STORAGE_PATH"]:
                 if user_id != "":
                     service.delete_temporary_user_data(user_id)
                 else:
